+++ HumbleKeysLibrarySettings.cs
@@ -8,27 +8,63 @@ using Playnite.SDK.Data;
 
 namespace HumbleKeys
 {
+    public enum RedemptionStoreType
+    {
+        None,       // 0
+        Source,     // 1
+        Tag,        // 2
+        Category,   // 3
+        Platform,   // 4
+    }
+
+    public enum TagMethodology
+    {
+        None,       // 0
+        Monthly,    // 1
+        All,        // 2
+    }
+
+    public enum UnredeemableMethodology
+    {
+        Tag,        // 0
+        Delete,     // 1
+    }
+
+    public class KeyInfo
+    {
+        public string Name { get; set; }        // Description used anywhere else needed, other than for Playnite's "Source" field
+        public string SourceName { get; set; }  // Exact match for Playnite's "Source" field, only used when Redemption Store is saved to Source
+        public Guid SourceId { get; set; }      // "Source" field GUID found in Playnite
+    }
+
     public class HumbleKeysLibrarySettings : ObservableObject, ISettings
     {
         private readonly HumbleKeysLibrary plugin;
-        private static ILogger logger = LogManager.GetLogger();
+        private static readonly ILogger logger = LogManager.GetLogger();
         private HumbleKeysLibrarySettings editingClone;
 
         public bool ConnectAccount { get; set; } = false;
         public bool IgnoreRedeemedKeys { get; set; } = false;
+        public bool AddKeyStatus { get; set; } = true;
+        public int RedemptionStore { get; set; } = (int)RedemptionStoreType.Source;
+        public bool AddLinks { get; set; } = true;
         public bool ImportChoiceKeys { get; set; } = false;
+        public int TagWithBundleName { get; set; } = (int)TagMethodology.None;
+        public int UnredeemableKeyHandling { get; set; } = (int)UnredeemableMethodology.Tag;
         public bool CacheEnabled { get; set; } = false;
-        public string CurrentTagMethodology { get; set; } = "none";
+        public bool AddPlatformNintendo { get; set; } = true;
+        public bool AddPlatformWindows { get; set; } = true;
 
-        public string CurrentUnredeemableMethodology { get; set; } = "tag";
-        
         [DontSerialize]
-        public List<string> keyTypeWhitelist = new List<string>() {
-            "gog",
-            "nintendo_direct",
-            "origin",
-            "origin_keyless",
-            "steam",
+        public Dictionary<string, KeyInfo> keyTypeWhitelist = new Dictionary<string, KeyInfo>
+        {
+            //["epic"] = new KeyInfo { Name = "Epic", SourceName = "Epic" },                  // This key type is valid so do we want to add it? I only have game dev asset keys at Epic myself right now but I assume this could be real games too
+            //["epic_keyless"] = new KeyInfo { Name = "Epic keyless", SourceName = "Epic" },  // Is this even a valid key type? I just guessed it might be because Humble mentions it has keyless Epic keys
+            ["gog"] = new KeyInfo { Name = "GOG", SourceName = "GOG" },
+            ["nintendo_direct"] = new KeyInfo { Name = "Nintendo", SourceName = "Nintendo" },
+            ["origin"] = new KeyInfo { Name = "EA", SourceName = "EA app" },
+            ["origin_keyless"] = new KeyInfo { Name = "EA keyless", SourceName = "EA app" },
+            ["steam"] = new KeyInfo { Name = "Steam", SourceName = "Steam" },
         };
 
         [DontSerialize]