+++ changelog.md
@@ -1,4 +1,15 @@
 ﻿## What's Changed
+# 0.3.9
+* IMPORTANT: Platform field is no longer used by default to store the Redemption Store, but now Source is
+* IMPORTANT: Your settings for "Create Tags for Bundle Names" and "Unredeemable key handling" will be reset to defaults due to changes in the internal data format for these options
+* Added setting: Checkbox to add Key Redemption status Tag (default enabled)
+* Added setting: Checkbox to "Add Humble & Steam links" (default enabled)
+* Added setting: Checkbox to add "Nintendo Switch" to Platform for all Nintendo keys
+* Added setting: Checkbox to add default of "PC (Windows)" to Platform for all other keys
+* Added setting: Dropdown to add Redemption Store (e.g. "Humble Key: Steam") to choice of field: Source (now default), Tag, Category, Platform (no longer default), or None (disabled)
+* Updated settings UI to be more compact and add the above features
+* Fixed a couple misc. bugs related to key redemption tags and tag methodology
+
 # 0.3.8
 * Restored plugin name
 * Restored plugin GUID to fix broken auto-update process from old versions, prevent duplicate old & new plugins installed at the same time, and ensure "Already installed" button works properly in Add-on Browser